third_party/WebKit/LayoutTests/editing/deleting/delete_map_area.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                '<span> test <map name=m><area coords="0,0,0,0" href=""></map> test| </span>',
            '</div>',
        ].join(''),
        selection => {
            for (var i = 0; i < 9; ++i)
                selection.document.execCommand('delete');
        },
        // BR is a placeholder.
        '<div contenteditable>|<br></div>'),
        'Delete command deletes MAP and AREA too.');
</script>
